// src/app/shared/models/email.ts
export class Email {
  user_auth_email: string;
  flag_primary: number = 0;
  stamp_created: any;
  stamp_created_by: {
    uuid: string;
    name_first: string;
    name_middle?: any;
    name_last: string;
  };
  stamp_updated: any;
  stamp_updated_by: {
    uuid: string;
    name_first: string;
    name_middle?: any;
    name_last: string;
  };
  stamp_email: 1584849407;
  stamp_email_by: {
    uuid: string;
    name_first: string;
    name_middle?: any;
    name_last: string;
  };
}